Update README.md
The Huffman algorithm assigns codes to each symbol so as to minimize the size of the message. The message is then encoded using this symbol-to-code mapping and transmitted to the receiver. For the receiver to decode the encoded message, it is necessary for the receiver to have some knowledge of the symbol-to-code mapping